Make inloadout search, organizer work
